base-files: for procd [able] services, start service on reload [if not running]
This is a continuation on change 6713694
That one changed the semantic from not doing restart on reload fail,
to defining default reload as restart.

But there is also the case where, if a service is not
running, reload could start it.

Of course, a reload hook could also implement this itself.
But, it feels like this would be a nice feature to have
for procd [able] services.

Still, if a service init script wants/needs to define
it's own `service_running` script, it can do that,
and it will be used instead of relying on procd [by default].

There are 2 base packages (system & qos-scripts) who's start & reload
calls do the same thing.
And it seems that for daemon-less init scripts, reload and start
do the same thing.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>