include: unpack.mk: impose PKG_BUILD_DIR directive
This is a behavioral build change, that should work on the basis that,
up-until now, PKG_BUILD_DIR would be used to specify the directory that
would be created by extracting from an archive.

Overriding PKG_BUILD_DIR, would not change the name of the top-level
directory from the archive, and thus whatever PKG_BUILD_DIR was specified,
the top-level dir (from the archive) would be extracted directly to
`<openwrt>/build_dir/<target-arch-libc>`. This would naturally fail when
trying to apply other build rules (patch, configure, compile, etc)

This means that if someone wants to override PKG_BUILD_DIR, he/she must
either, also override the PKG_UNPACK rule, to make sure that the folder
is unpack-ed to the new value of PKG_BUILD_DIR.

This can also be done by doing a untar + `--strip-components=1`, which
strips the top-level dir (when extracting) and just dumping the files into
the new folder.

This (using `--strip-components=1`) is also done in the buildroot system,
so this is not re-inventing the wheel.

The unzip command (however) does not have a `--strip-components` equivalent
command, so a move must be performed.
At this point in time, there are only a few .zip packages:
1. wolfssl in openwrt core (this can be switched to tar.gz)
2. ixp4xx-microcode in openwrt core
3. lua-lsqlite3, django-simple-captcha, python-gmpy2, opencv, rxtx,
   vncrepeater & espeak in packages feed.

The current model, also assumed that the directories would get
untarred/unzipped to $(PKG_BUILD_DIR)/.. , which would (naturally) just
use the top-level dir from the archive. Which is why, there shouldn't be
any breakage with this patch, but rather more of an enhancement, which
would allow to remove some explicit PKG_UNPACK rules (especially in
packages feed), and maybe also a few explicit PKG_BUILD_DIR vars.

This patch was tested with all zip files.

Naturally, all tar.XX files also build fine.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>